Moved staff allocation tests into a sub-package for better organisation.
